[ADDED] finish ControlCameraOp, rotate, translate and zoom state

+++ easy3d/src/easy3d/ViewFrustum.h
@@ -9,7 +9,7 @@ namespace e3d
 class ViewFrustum
 {
 public:
-	static sm::ivec2 TransPos3ProjectToScreen(const sm::vec3& proj, float screen_width, float screen_height);
+	static sm::vec2 TransPos3ProjectToScreen(const sm::vec3& proj, float screen_width, float screen_height);
 	static sm::vec3 TransPos3ScreenToProject(const sm::ivec2& scr, float proj_z, float screen_width, float screen_height);
 
 	static sm::vec3 TransPos3ScreenToDir(const sm::ivec2& screen, float screen_width, float screen_height);
